// src/generated/api/orders/models/listAdminOrdersStatusGroup.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type ListAdminOrdersStatusGroup =
  (typeof ListAdminOrdersStatusGroup)[keyof typeof ListAdminOrdersStatusGroup];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ListAdminOrdersStatusGroup = {
  PENDING_CONFIRMATION: 'PENDING_CONFIRMATION',
  CONFIRMED: 'CONFIRMED',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
} as const;